legend in CompareWithTGenPhaseSpace.cu

+++ mcbooster/GTypes.h
@@ -47,7 +47,7 @@ typedef unsigned char GByte_t;      ///< Byte (8 bits) (unsigned char)
 typedef long long GLong64_t; ///< Portable signed long integer 8 bytes
 typedef unsigned long long GULong64_t; ///< Portable unsigned long integer 8 bytes
 #ifdef FP_SINGLE
-typedef double GReal_t;///< Double 8 bytes or float 4 bytes
+typedef float GReal_t;///< Double 8 bytes or float 4 bytes
 #else
 typedef double GReal_t;///< Double 8 bytes or float 4 bytes
 #endif
